move datasetbuilder and test files to utils
